feat: implement OG image edge handler
