+++ site/vercel.json
@@ -9,5 +9,13 @@
       "source": "/robots.txt",
       "headers": [{ "key": "Cache-Control", "value": "public, max-age=3600" }]
     }
-  ]
+  ],
+  "functions": {
+    "api/og/*.js": {
+      "runtime": "edge"
+    },
+    "api/og/*.ts": {
+      "runtime": "edge"
+    }
+  }
 }
